Some Ala42 eagle eye catch!

+++ Libraries/AQ_BarometricSensor/BarometricSensor.h
@@ -24,14 +24,14 @@
 #include "Arduino.h"
 #include "GlobalDefined.h"
 
-double baroAltitude      = 0.0; 
-double baroRawAltitude   = 0.0;
+float baroAltitude      = 0.0; 
+float baroRawAltitude   = 0.0;
 float baroGroundAltitude = 0.0;
 float baroSmoothFactor   = 0.02;
 
-double baroAltitudeSampleSum = 0.0;
+float baroAltitudeSampleSum = 0.0;
 byte baroAltitudeSampleSumCount = 0;
-double oversampledBaroAltitude = 0.0;
+float oversampledBaroAltitude = 0.0;
   
 // **********************************************************************
 // The following function calls must be defined inside any new subclasses
